Migrate rating methods to ratings.py
ratings.py now depends on model.py, but it makes sense to keep all the "rating" logic within that file.
